Refactor export tests along the DatasetExport/ImageDatasetExport seam
Move the format-contract tests (COCO/YOLO/Pascal/captions output correctness,
category sorting, mask handling, path defaults) from TestImageDatasetExport into
a new TestDatasetExport, driving DatasetExport directly with the image mapping as
substrate. These assert behavior shared by every sample type, so they belong to
the base.

Slim test_image_dataset_export.py to the image-specific delta: unit tests for the
image_sample_to_image mapping (absolute path for COCO, relative name for
YOLO/Pascal) plus one smoke that ImageDataset.export() binds it.

No behavior change; coverage preserved. An upcoming VideoFrameDatasetExport only
needs its own small mapping tests, not a copy of the format suite.
